requirements.txt
autopep8==1.6.0
bitarray-ph4==1.9.1
pycodestyle==2.8.0
toml==0.10.2
